refactor: move common method of subclasses to the parent class
