+++ scripts/cloudfunctions/recommender-checker/localpackage/recommender/compute/idle_resource.py
@@ -1,7 +1,3 @@
-import os
-from google.cloud.asset_v1.types.assets import ResourceSearchResult
-from proto.marshal.collections.repeated import RepeatedComposite
-from typing import List
 from ..recommender import Recommender
 
 
@@ -24,28 +20,3 @@ class VMIdleResourceRecommender(Recommender):
                 payload = self._generate_slack_payload(project_name, recommendation)
                 print(payload)
                 self._post_slack_message(payload)
-
-    def _search_assets(self) -> List[ResourceSearchResult]:
-        asset_list = []
-        page_token = ''
-        while True:
-            response = self.asset_client.search_all_resources(
-                request={
-                    "scope": f"organizations/{os.environ['ORGANIZATION_ID']}",
-                    "asset_types": [self.asset_type],
-                    "page_token": page_token
-                }
-            )
-            asset_list += response.results
-            page_token = response.next_page_token
-            if not page_token:
-                break
-        return asset_list
-
-    def _list_recommendations(self, project_number: str, zone: str) -> RepeatedComposite:
-        response = self.recommender_client.list_recommendations(
-            request={
-                "parent": f"projects/{project_number}/locations/{zone}/recommenders/{self.recommender_id}"
-            }
-        )
-        return response.recommendations
