+++ scripts/cloudfunctions/recommender-checker/localpackage/recommender/recommender.py
@@ -2,8 +2,10 @@ import json
 import os
 from google.cloud import asset_v1
 from google.cloud import recommender_v1
+from google.cloud.asset_v1.types.assets import ResourceSearchResult
 from google.cloud.recommender_v1.types.recommendation import Recommendation
-from typing import TypedDict
+from proto.marshal.collections.repeated import RepeatedComposite
+from typing import List, TypedDict
 from urllib.request import Request, urlopen
 
 
@@ -19,11 +21,30 @@ class Recommender():
         self.asset_type = asset_type
         self.ignore_descriptions = ignore_descriptions
 
-    def _search_assets(self):
-        raise NotImplementedError
+    def _search_assets(self) -> List[ResourceSearchResult]:
+        asset_list = []
+        page_token = ''
+        while True:
+            response = self.asset_client.search_all_resources(
+                request={
+                    "scope": f"organizations/{os.environ['ORGANIZATION_ID']}",
+                    "asset_types": [self.asset_type],
+                    "page_token": page_token
+                }
+            )
+            asset_list += response.results
+            page_token = response.next_page_token
+            if not page_token:
+                break
+        return asset_list
 
-    def _list_recommendations(self):
-        raise NotImplementedError
+    def _list_recommendations(self, project_number: str, zone: str) -> RepeatedComposite:
+        response = self.recommender_client.list_recommendations(
+            request={
+                "parent": f"projects/{project_number}/locations/{zone}/recommenders/{self.recommender_id}"
+            }
+        )
+        return response.recommendations
 
     def _generate_slack_payload(self, project: str, recommendation: Recommendation) -> SlackPayload:
         _id = self.recommender_id.strip('google.')
